qurio-chatbot/requirements.txt
Flask
Flask-Cors
firebase-admin
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
APScheduler
requests
python-dotenv
gunicorn
